fix: resolve clang-cl build errors
- CombatBehaviorStack: add missing pad0C field initializer
- MotionDataContainer: replace MSVC-specific __m128 member access with portable SSE intrinsics
- CombatBehaviorTreeNodeObject: remove explicit instantiations that conflict with specializations
- CombatBehaviorChildSelector/FallbackSelector: fix signed/unsigned comparison
- hkArray: fix signed/unsigned comparison in assert
- Projectile: add missing field initializer for ProjectileRot
- ChildSelectorBase: move inline definitions to .cpp
- TESObjectREFR: use SKYRIM_REL_VR_VIRTUAL for Unk_84

+++ include/RE/C/CombatBehaviorChildSelector.h
@@ -25,7 +25,7 @@ namespace RE
 		{
 			auto cur_thread = GetStaticTLSData()->combatAIThread;
 			auto ind = Selector::SelectChild(const_cast<CombatBehaviorTreeNode*>(cur_thread->currentNode), 0xFFFFFFFF);
-			if (ind == -1) {
+			if (ind == static_cast<uint32_t>(-1)) {
 				cur_thread->SetFailed(true);
 				cur_thread->Ascend();
 			} else {
